// src/api/__generated__/wallet/model/cursorPageResponseInviteRecordItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { InviteRecordItemResponse } from './inviteRecordItemResponse';

/**
 * 游标分页响应
 */
export interface CursorPageResponseInviteRecordItemResponse {
  /** 每页数量 */
  pageSize?: number;
  /** 下一页游标，-1 表示没有更多数据 */
  mark?: string;
  /** 是否还有更多数据 */
  hasMore?: boolean;
  /** 分页数据 */
  list?: InviteRecordItemResponse[];
}