// src/api/__generated__/wallet/model/inviteRecordItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 邀请记录
 */
export interface InviteRecordItemResponse {
  /** 被邀请用户ID */
  inviteeUserId?: string;
  /** 被邀请用户昵称 */
  nickname?: string;
  /** 被邀请用户头像URL */
  avatarUrl?: string;
  /** 被邀请用户注册时间 */
  createdAt?: string;
  /** 邀请绑定时间 */
  bindAt?: string;
  /** 是否有效用户：1-有效，0-待激活 */
  isValid?: number;
}